+++ server/controllers/invoice/invoiceArtisan/CreateArtisanInvoiceController.js
@@ -1,8 +1,8 @@
-const db = require("../../data/index.js");
+const db = require("../../../data/index.js");
 const { Invoice, InvoiceItem, Company, Activity, Measure, Project, Task, Artisan, WorkItem, DefaultPricing } = db;
-const { createArtisanInvoicePDF } = require("../../utils/pdfGenerator.js");
-const { sendInvoiceEmail } = require("../../utils/invoiceEmailService.js");
-const { sequelize } = require("../../data/index.js");
+const { createArtisanInvoicePDF } = require("../../../utils/pdfGenerator.js");
+const { sendInvoiceEmail } = require("../../../utils/invoiceEmailService.js");
+const { sequelize } = require("../../../data/index.js");
 
 const getWeekNumber = date => {
   const d = new Date(date);
