+++ server/routes/invoiceArtisanRoutes.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const { createArtisanInvoice, getArtisanInvoicePDF, updateArtisanInvoice, deleteArtisanInvoice } = require("../controllers/invoiceArtisan/CreateArtisanInvoiceController");
+const { createArtisanInvoice, getArtisanInvoicePDF, updateArtisanInvoice, deleteArtisanInvoice } = require("../controllers/invoice/invoiceArtisan/CreateArtisanInvoiceController");
 // const { getArtisanInvoiceById } = require("../controllers/invoiceArtisan/GetArtisanInvoiceByIdController");
 // const { getPDFArtisanInvoiceById } = require("../controllers/invoiceArtisan/GetPDFArtisanInvoiceByIdController");
 // const { updateArtisanInvoiceStatus } = require("../controllers/invoiceArtisan/UpdateArtisanInvoiceStatusController");
