+++ server/routes/invoiceClientRoutes.js
@@ -1,12 +1,12 @@
 const express = require("express");
 const router = express.Router();
-const { createClientInvoice } = require("../controllers/invoiceClient/CreateClientInvoiceController");
-const { editClientInvoice } = require("../controllers/invoiceClient/EditClientInvoiceController");
-const { getAllClientInvoices } = require("../controllers/invoiceClient/GetAllClientInvoicesController");
-const { getClientInvoiceById } = require("../controllers/invoiceClient/GetClientInvoiceByIdController");
-const { getPDFClientInvoiceById } = require("../controllers/invoiceClient/GetPDFClientInvoiceByIdController");
-const { updateClientInvoiceStatus } = require("../controllers/invoiceClient/UpdateClientInvoiceStatusController");
-const { deleteClientInvoice } = require("../controllers/invoiceClient/DeleteClientInvoiceController");
+const { createClientInvoice } = require("../controllers/invoice/invoiceClient/CreateClientInvoiceController");
+const { editClientInvoice } = require("../controllers/invoice/invoiceClient/EditClientInvoiceController");
+const { getAllClientInvoices } = require("../controllers/invoice/invoiceClient/GetAllClientInvoicesController");
+const { getClientInvoiceById } = require("../controllers/invoice/invoiceClient/GetClientInvoiceByIdController");
+const { getPDFClientInvoiceById } = require("../controllers/invoice/invoiceClient/GetPDFClientInvoiceByIdController");
+const { updateClientInvoiceStatus } = require("../controllers/invoice/invoiceClient/UpdateClientInvoiceStatusController");
+const { deleteClientInvoice } = require("../controllers/invoice/invoiceClient/DeleteClientInvoiceController");
 
 router.post("/create", createClientInvoice);
 router.put("/:id/edit", editClientInvoice);
